+++ src/Build-LinuxClient.ps1
@@ -4,7 +4,7 @@
 param(
     [Parameter()]
     [ValidateNotNullOrEmpty()]
-    [string]$Version = '0.1.4',
+    [string]$Version = '0.1.5',
 
     [Parameter()]
     [ValidateNotNullOrEmpty()]